This is my final code.

+++ BadMath/Source.cpp
@@ -2,15 +2,15 @@
 #include <iomanip>
 using namespace std;
 
-float average(int i1, int i2, int i3)
+float average(float i1, float i2, float i3)
 {
 	float sum = i1 + i2 + i3;
-	return  sum  / 3;
+	return  sum/3;
 }
 
 int main()
 {
-	int n1 =0 , n2 = 0, n3 = 0;
+	float n1 =0 , n2 = 0, n3 = 0;
 	cout << "This program calculates the average of three numbers." << endl;
 	cout << "First number: "; cin >> n1;
 	cout << "Second number: "; cin >> n2;
